+++ lib/brando/live_preview.ex
@@ -24,17 +24,20 @@ defmodule Brando.LivePreview do
 
   def initialize(schema, entry, key, prop) do
     preview_module = get_preview_module()
-    cache_key = build_cache_key(:erlang.system_time())
-    schema_module = Module.concat([schema])
-    entry_struct = Brando.Utils.stringy_struct(schema_module, entry)
-    wrapper_html = preview_module.render(schema_module, entry_struct, key, prop, cache_key)
 
-    if Cachex.get(:cache, cache_key) == {:ok, nil} do
-      Brando.LivePreview.store_cache(cache_key, wrapper_html)
+    if function_exported?(preview_module, :render, 5) do
+      cache_key = build_cache_key(:erlang.system_time())
+      schema_module = Module.concat([schema])
+      entry_struct = Brando.Utils.stringy_struct(schema_module, entry)
+      wrapper_html = preview_module.render(schema_module, entry_struct, key, prop, cache_key)
+
+      if Cachex.get(:cache, cache_key) == {:ok, nil} do
+        Brando.LivePreview.store_cache(cache_key, wrapper_html)
+      end
+
+      Brando.endpoint().broadcast("live_preview:#{cache_key}", "update", %{html: wrapper_html})
+      cache_key
     end
-
-    Brando.endpoint().broadcast("live_preview:#{cache_key}", "update", %{html: wrapper_html})
-    cache_key
   end
 
   def update(schema, entry, key, prop, cache_key) do