+++ priv/templates/brando.upgrade/migrations/brando_21_add_meta_image_to_pages.exs
@@ -1,0 +1,10 @@
+defmodule Brando.Repo.Migrations.AddMetaImageToPages do
+  use Ecto.Migration
+  use Brando.Sequence.Migration
+
+  def change do
+    alter table(:pages_pages) do
+      add :meta_image, :jsonb
+    end
+  end
+end